# backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: CoffeeChain Trace API
  description: |
    REST API for the CoffeeChain coffee supply-chain traceability backend.
    All authenticated endpoints require a Bearer JWT obtained from POST /api/auth/login.
  version: 1.0.0
  contact:
    name: CoffeeChain Backend Unit-2
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/process:
    post:
      tags: [Processor]
      summary: Create processed batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcessedBatchRequest"
      responses:
        "200":
          description: Created batch from ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/process/{id}/status:
    patch:
      tags: [Processor]
      summary: Update processed batch status
      parameters:
        - $ref: "#/components/parameters/batchId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Updated batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchView"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─────────────────────────────────────────────
  # Roaster
  # ─────────────────────────────────────────────
...
  /api/transfer/accept/{id}:
    post:
      tags: [Packager]
      summary: Accept transfer (SBE AND — Org1 + Org2 must endorse)
      parameters:
        - $ref: "#/components/parameters/batchId"
      responses:
        "200":
          description: Accepted batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/package:
    post:
      tags: [Packager]
      summary: Create packaged batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackagedBatchRequest"
      responses:
        "200":
          description: Created batch from ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/package/{id}/qr:
    get:
      tags: [Packager]
      summary: Get QR code image for a packaged batch
      parameters:
        - $ref: "#/components/parameters/batchId"
      responses:
        "200":
          description: PNG QR code image
          content:
            image/png:
              schema:
                type: string
                format: binary
        "501":
          description: QrCodeService not yet integrated (Unit-3 pending)

  # ─────────────────────────────────────────────
  # Retailer
  # ─────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    batchId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Internal batch UUID on the ledger
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but insufficient role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
...
    BatchView:
      type: object
      additionalProperties: true
      description: Raw JSON returned from the Fabric chaincode (schema mirrors Batch.java on ledger)
...
    CreateProcessedBatchRequest:
      type: object
      required:
        [
          parentBatchId,
          processingMethod,
          startDate,
          endDate,
          facilityName,
          weightKg,
        ]
      properties:
        parentBatchId:
          type: string
        processingMethod:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        facilityName:
          type: string
        weightKg:
          type: string
...
    CreatePackagedBatchRequest:
      type: object
      required:
        [parentBatchId, packageWeight, packageDate, expiryDate, packageCount]
      properties:
        parentBatchId:
          type: string
        packageWeight:
          type: string
        packageDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        packageCount:
          type: string
...
    UpdateStatusRequest:
      type: object
      required: [newStatus]
      properties:
        newStatus:
          type: string
          enum:
            [
              CREATED,
              IN_PROCESS,
              COMPLETED,
              TRANSFER_PENDING,
              TRANSFERRED,
              IN_STOCK,
              SOLD,
            ]
        finalWeightKg:
          type: string
          description: Final weight (required when newStatus=COMPLETED for harvest/processed/roast)
        roastDurationMinutes:
          type: string
          description: Roast duration in minutes (required when newStatus=COMPLETED for roast)